Extract settings utility functions into settingsUtils.ts
